app/requirements.txt
gradio==4.44.1
openai_whisper==20231117
pydub==0.25.1
